Refactor and clean up .rubocop_todo.yml
