+++ .rubocop_todo.yml
@@ -2,13 +2,13 @@ inherit_from: .rubocop.yml
 
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-02-19 14:32:34 +0000 using RuboCop version 0.64.0.
+# on 2019-02-19 15:47:19 +0000 using RuboCop version 0.64.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 2
+# Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 244
+  Max: 111
